requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-telegram-bot==21.9
aiohttp==3.11.11
aiosqlite==0.20.0
yt-dlp>=2024.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
psutil>=5.9.0
pytest>=8.0.0
pytest-asyncio==0.23.7
trio>=0.25.0
boto3>=1.34.0
aioboto3>=12.3.0